refactor: structure reorganization, model rename
