--- iOS-FakeNFT-Extended/Profile/ProfileUIModel.swift
@@ -1,10 +1,0 @@
-import Foundation
-
-struct ProfileUIModel {
-    let avatarURL: URL?
-    let name: String
-    let description: String
-    let websiteURL: URL?
-    let nftsCount: String
-    let likesCount: String
-}
